v0.2.0 Push - Added memory agent and memory tools. Added longterm memory decision for each prompt. Reorganized tools. Updated versioning.
