+++ README.md
@@ -1,7 +1,6 @@
-# Alfred v0.1.3 - README
+# Alfred v0.2.0 - README
 
-Alfred is a lightweight AI assistant project powered by a FastAPI backend, a React frontend, OpenAI GPT models for the main agent,  
-and a LLaMA 3 local model running via Ollama for secondary agents.
+Alfred is a lightweight AI assistant project powered by a FastAPI backend, a React frontend, OpenAI GPT models for the main agent, and a LLaMA 3 local model running via Ollama for secondary agents.
 
 ## Demo
