+++ backend/nodes/tool_node.py
@@ -1,4 +1,4 @@
-# alfred_v0.1.3/backend/nodes/tool_node.py
+# alfred/backend/nodes/tool_node.py
 
 def tool_node(state):
     return state
