+++ backend/utils/tools/location_date_tools.py
@@ -1,0 +1,32 @@
+# alfred/backend/utils/tools/location_date_tools.py
+
+from datetime import datetime
+import requests
+import os
+
+DEBUG = False
+
+def get_current_date(query: str) -> str:
+    """
+    Use this tool when you need the current date and/or time to complete a task.
+    """
+    if DEBUG:
+        print(f"ALFRED USING GET CURRENT TIME TOOL")
+    return datetime.now().strftime("%Y-%m-%d %H:%M")
+
+def get_lat_lon(city: str):
+    """
+    Use this tool to turn a location into a lat/lon
+    """
+    print(f"WEATHER AGENT USING GET LAT LON TOOL FOR {city}")
+    url = f"https://api.opencagedata.com/geocode/v1/json?q={city}&key={os.getenv('GEOCODE_API')}"
+    response = requests.get(url)
+    data = response.json()
+    if data['results']:
+        lat = data['results'][0]['geometry']['lat']
+        lon = data['results'][0]['geometry']['lng']
+        if DEBUG:
+            print(f"LAT LON: {lat},{lon}")
+        return lat, lon
+    else:
+        return None, None
