+++ frontend/src/index.css
@@ -1,4 +1,4 @@
-/* alfred_v0.1.2/alfred-vite/src/index.css */
+/* alfred/frontend/src/index.css */
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
