+++ version.py
@@ -1,0 +1,3 @@
+# alfred/version.py
+
+__version__ = "0.2.0"
